Phylogenetic: Construct phylogeny rules
Migrate rules for constructing phylogenetic trees to the construct_phylogeny.smk snakemake file to follow the pathogen-repo-guide with some adjustments:

* Add log and benchmark directives
* Explicitly set metadata-id-columns

Note: Some of 54 trees were failing at the refine step due to negative rate, so an earlier commit changed these to divergence trees. Later on, we can revisit which ones will and will not resolve with best (least-squares) rooting.